ML/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
cycler==0.11.0
Flask==2.2.5
fonttools==4.38.0
idna==3.4
kiwisolver==1.4.5
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.8.1.78
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pip==23.2.1
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scipy==1.7.3
seaborn==0.12.2
setuptools==40.6.2
six==1.16.0
supervision==0.3.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
typing_extensions==4.7.1
ultralytics==8.0.145
urllib3==1.26.6
